RF_11.001.01 | MainPage2Test
Description:
Moved all view description related tests to GlobalViewTest class.
